Upgrade class
Ii put in an upgrade class. It is currently unimplemented though. there is no way to increase the values.
