+++ Scenes/hud.tscn
@@ -203,9 +203,8 @@ script = ExtResource("5_4f5kr")
 
 [node name="Button" type="Button" parent="."]
 offset_left = 1018.0
-offset_top = 634.0
 offset_right = 1150.0
-offset_bottom = 648.0
+offset_bottom = 14.0
 mouse_filter = 2
 flat = true
 script = ExtResource("5_4f5kr")
