+++ Scenes/main.tscn
@@ -86,17 +86,11 @@ script = ExtResource("4_82xsv")
 [node name="Player-1" parent="." node_paths=PackedStringArray("target") instance=ExtResource("6_3pmx8")]
 position = Vector2(896, 353)
 target = NodePath("../Player-2")
-brute = null
-ranged = null
-healer = null
 
 [node name="Player-2" parent="." node_paths=PackedStringArray("target") instance=ExtResource("6_3pmx8")]
 position = Vector2(112, 346)
 player1 = false
 target = NodePath("../Player-1")
-brute = null
-ranged = null
-healer = null
 
 [node name="NavHandler" type="Node2D" parent="."]
 script = ExtResource("10_wsu2k")
