+++ Scripts/TroopUpgrade.cs
@@ -1,0 +1,11 @@
+using Godot;
+using System;
+
+public partial class TroopUpgrade : Node
+{
+	public GameManager.Towers troopType;
+	public int upgradeLevel;
+	public int speedLevel;
+	public int healthLevel;
+	public int damageLevel;
+}
