--- Troops/Scripts/BaseTroop.cs.uid
@@ -1,1 +1,0 @@
-uid://dy2ppg2hig8q6
