+++ project.godot
@@ -11,7 +11,7 @@ config_version=5
 [application]
 
 config/name="Forged in Time"
-run/main_scene="uid://cfv3xbll1nk1r"
+run/main_scene="uid://bywcqda0sombh"
 config/features=PackedStringArray("4.4", "C#", "GL Compatibility")
 config/icon="res://Assets/icon.svg"
